refactor qmc5883: use driver base class
